Update patch set 2
Patch Set 2:

(3 comments)

This is a fairly large change and I can only skim for obvious mistakes—of which none are evident. But microhttpd does feel like a future security problem in the making. Are you sure that you can't use a Go binary or something safer for this?

Patch-set: 2
Reviewer: Gerrit User 1019036 <1019036@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0