Remove GCM cipher mode requirements for fallback ciphersuites
SecureTransport of MacOS/iOS does not support GCM cipher mode with
ECDHE-ECDSA/ECDHE-RSA key-exchange algorithms even though they seem
to claim that they do. This is supposedly fixed in MacOS 10.11 but
that one is not out yet.

Using "meta" suite from BoringSSL to support strong encryption
algorithms with a bit more leeway, not just 4 we used to hardcode.

BUG: 24468826
Change-Id: Ie62a9b11584b6745848090888b7d9e1fe50e90ee